Cleanup and separate files
Split the file to have module specific procedures in upf.go
